app/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
kafka-python==2.0.2
python-dotenv==1.0.1
pydantic==2.6.4
httpx==0.27.0
